refactor: 게시글 수집 및 알림 시스템 객체 명, 패키지 위치 변경
