--- src/main/java/com/flytrap/rssreader/api/alert/business/event/NewPostAlertEventHolder.java
@@ -1,7 +1,0 @@
-package com.flytrap.rssreader.api.alert.business.event;
-
-import com.flytrap.rssreader.global.event.EventHolder;
-import lombok.Getter;
-
-public record NewPostAlertEventHolder(@Getter NewPostAlertEventParam value) implements EventHolder<NewPostAlertEventParam> {
-}
