+++ src/main/java/com/flytrap/rssreader/api/alert/business/event/NewPostAlertEventListener.java
@@ -1,7 +1,7 @@
 package com.flytrap.rssreader.api.alert.business.event;
 
-import com.flytrap.rssreader.api.alert.business.service.AlertService;
 import com.flytrap.rssreader.api.alert.domain.Alert;
+import com.flytrap.rssreader.api.alert.infrastructure.system.AlertSendSystem;
 import com.flytrap.rssreader.api.folder.business.service.FolderReadService;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
@@ -11,25 +11,22 @@ import org.springframework.stereotype.Component;
 
 @Component
 @RequiredArgsConstructor
-public class NewPostAlertEventHandler {
+public class NewPostAlertEventListener {
 
-    private final AlertService alertService;
+    private final AlertSendSystem alertSendSystem;
     private final FolderReadService folderReadService;
 
     @Async
-    @EventListener(NewPostAlertEventHolder.class)
-    public void handle(NewPostAlertEventHolder event) {
-        NewPostAlertEventParam param = event.value();
-
-        List<Alert> alerts = alertService.getAlertListBySubscribe(param.subscribe().getId());
+    @EventListener(NewPostAlertEvent.class)
+    public void handle(NewPostAlertEvent event) {
+        List<Alert> alerts = alertSendSystem.getAlertListBySubscribe(event.subscribe().getId());
         if (!alerts.isEmpty()) {
             alerts.forEach(alert ->
-                alertService.sendAlertToPlatform(
+                alertSendSystem.sendAlertToPlatform(
                     folderReadService.findById(alert.getFolderId()).getName(),
                     alert.getWebhookUrl(),
-                    param.posts()
+                    event.posts()
                 ));
         }
     }
 }
-
