+++ src/main/java/com/flytrap/rssreader/api/alert/business/service/AlertService.java
@@ -5,16 +5,12 @@ import com.flytrap.rssreader.api.alert.domain.Alert;
 import com.flytrap.rssreader.api.alert.domain.AlertCreate;
 import com.flytrap.rssreader.api.alert.domain.AlertId;
 import com.flytrap.rssreader.api.alert.domain.AlertPlatform;
-import com.flytrap.rssreader.api.alert.infrastructure.entity.AlertEntity;
-import com.flytrap.rssreader.api.alert.infrastructure.external.AlertSender;
 import com.flytrap.rssreader.api.alert.infrastructure.implement.AlertCommand;
 import com.flytrap.rssreader.api.alert.infrastructure.implement.AlertQuery;
 import com.flytrap.rssreader.api.alert.infrastructure.implement.AlertValidator;
-import com.flytrap.rssreader.api.alert.infrastructure.repository.AlertEntityJpaRepository;
 import com.flytrap.rssreader.api.folder.domain.FolderDomain;
 import com.flytrap.rssreader.api.folder.domain.FolderId;
 import com.flytrap.rssreader.api.folder.infrastructure.implementatioin.FolderValidator;
-import com.flytrap.rssreader.api.post.infrastructure.entity.PostEntity;
 import com.flytrap.rssreader.global.exception.domain.ForbiddenAccessFolderException;
 import com.flytrap.rssreader.global.exception.domain.NoSuchDomainException;
 import java.util.List;
@@ -24,9 +20,6 @@ import org.springframework.stereotype.Service;
 @Service
 @RequiredArgsConstructor
 public class AlertService {
-
-    private final AlertEntityJpaRepository alertRepository;
-    private final List<AlertSender> alertSenders;
 
     private final AlertQuery alertQuery;
     private final AlertCommand alertCommand;
@@ -60,23 +53,6 @@ public class AlertService {
             throw new NoSuchDomainException(Alert.class);
 
         alertCommand.delete(alertId);
-    }
-
-    public void sendAlertToPlatform(String folderName, String webhookUrl, List<PostEntity> posts) {
-        AlertPlatform alertPlatform = AlertPlatform.parseWebhookUrl(webhookUrl);
-
-        for (AlertSender alertSender : alertSenders) {
-            if (alertSender.isSupport(alertPlatform)) {
-                alertSender.sendAlert(folderName, webhookUrl, posts);
-            }
-        }
-    }
-
-    public List<Alert> getAlertListBySubscribe(Long subscribeId) {
-        return alertRepository.findAlertsBySubscribeId(subscribeId)
-                .stream()
-                .map(AlertEntity::toReadOnly)
-                .toList();
     }
 
 }
