+++ src/main/java/com/flytrap/rssreader/api/post/business/event/PostCollectEventListener.java
@@ -1,6 +1,6 @@
 package com.flytrap.rssreader.api.post.business.event;
 
-import com.flytrap.rssreader.api.post.business.service.collect.PostCollectService;
+import com.flytrap.rssreader.api.post.infrastructure.system.PostCollectSystem;
 import lombok.RequiredArgsConstructor;
 import org.springframework.context.event.EventListener;
 import org.springframework.scheduling.annotation.Async;
@@ -10,11 +10,11 @@ import org.springframework.stereotype.Component;
 @RequiredArgsConstructor
 public class PostCollectEventListener {
 
-    private final PostCollectService postCollectService;
+    private final PostCollectSystem postCollectSystem;
 
     @Async
     @EventListener(PostCollectEvent.class)
     public void handle(PostCollectEvent event) {
-        postCollectService.enqueueHighPrioritySubscription(event.subscribeEntity());
+        postCollectSystem.enqueueHighPrioritySubscription(event.subscribeEntity());
     }
 }
