--- src/main/java/com/flytrap/rssreader/api/post/business/service/collect/CollectPriority.java
@@ -1,5 +1,0 @@
-package com.flytrap.rssreader.api.post.business.service.collect;
-
-public enum CollectPriority {
-    HIGH, LOW
-}
