+++ src/main/java/com/flytrap/rssreader/api/post/infrastructure/system/CollectPriority.java
@@ -1,0 +1,5 @@
+package com.flytrap.rssreader.api.post.infrastructure.system;
+
+public enum CollectPriority {
+    HIGH, LOW
+}
